+++ lib/flutter/gdi/animation.dart
@@ -17,13 +17,54 @@ class _AnimationFrame
   const _AnimationFrame(this.imageInfo, this.duration, this.image );
 }
 
-/// Represent a simple image animation owned by [WxAnimationCtrl]
+/// Represents a simple image animation owned by [WxAnimationCtrl].
 /// 
-/// Not to be mixed up with [WxUIAnimation]
+/// You can used the derived class [WxGraphicsAnimation] to draw 
+/// the animation frames directly into a window using a [WxGraphicsContext].
+/// 
+/// Not to be mixed up with [WxUIAnimation] which controls animations 
+/// onscreen.
+///
+/// This is how to retrieve the individual frames of an animation and
+/// convert them to [WxBitmap]s. You need to call [load] as the animation
+/// is loaded asynchronously.
+/// 
+/// ```dart
+///  String path = wxGetStandardPaths().getResourcesDir();
+///  // Add forward or backward slash
+///  if (wxIsMSW() && !wxUsesFlutter()) {
+///    path += "\\throbber.gif";
+///  } else {
+///    path += "/throbber.gif";
+///  }
+/// 
+///  final List<WxBitmap> frames = [];
+///  final animation = WxAnimation(path);
+///  animation.load().then( (_) {
+///    if (!animation.isOk()) {
+///      wxLogError( "animation did not load" );
+///      return;
+///    }
+///    for (int i = 0; i < animation.getFrameCount(); i++ ) {
+///      final image = animation.getFrame(i);
+///      if (image != null) {
+///        if (!image.hasAlpha()) {
+///         // this will convert from mask to alpha which is
+///         // currently required in wxDart Native
+///         image.initAlpha();
+///      }
+///      frames.add( WxBitmap.fromImage(image) );
+///    }
+///    refresh();
+///  } );
+/// ```
 
 class WxAnimation extends WxObject {
   /// Creates an animation from a GIF image. Supports WebP and ANI 
-  /// formats on wxDart Native as well.
+  /// formats on wxDart Native (Windows, macOS) as well.
+  /// 
+  /// Note that the wxDart Native implementation does not accept animations
+  /// that are incorrectly encoded as GIF87a files, only GIF89a files.
   /// 
   /// [path] is relative to [WxStandardPaths.getResourcesDir]
   WxAnimation( String path ) {
@@ -85,7 +126,8 @@ class WxAnimation extends WxObject {
     return WxSize( _list.first.imageInfo.image.width, _list.first.imageInfo.image.height );
   }
 
-  /// Load the animation. Returns the future when fully loaded.
+  /// Loads the animation asynchronously. Call this if you need to access the
+  /// individual frame as [WxImage]s.
   Future<void> load() async
   {
     if (_isOk) {
@@ -110,7 +152,7 @@ class WxAnimation extends WxObject {
         _AnimationFrame(
           ImageInfo(image: frameInfo.image),
           frameInfo.duration,
-          byteData!.lengthInBytes == sizeRGBA 
+          byteData.lengthInBytes == sizeRGBA 
             ? WxImage.fromRGBA( uiImage.width, uiImage.height, byteData.buffer.asUint8List() )
             : null
         ),